backend/requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-bcrypt
flask-jwt-extended
flask-cors
flask-marshmallow
marshmallow-sqlalchemy
python-dotenv
google-generativeai
psycopg2-binary
feedparser
beautifulsoup4
lxml
requests
deep-translator
openpyxl
gunicorn